+++ package.json
@@ -9,8 +9,6 @@
   },
   "dependencies": {
     "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
-    "connect-history-api-fallback": "^1.5.0",
     "path": "^0.12.7",
     "vue": "^2.5.17",
     "vue-axios": "^2.1.4",
@@ -25,7 +23,6 @@
     "eslint": "^5.8.0",
     "eslint-plugin-vue": "^5.0.0-0",
     "express": "^4.16.4",
-    "serve-static": "^1.13.2",
     "vue-template-compiler": "^2.5.17"
   },
   "eslintConfig": {
