+++ server.js
@@ -1,0 +1,16 @@
+const express = require('express');
+const path = require('path');
+const app = express();
+app.use(express.json());
+app.use(express.urlencoded({
+    extended: true
+}));
+
+app.use(express.static('./dist'))
+app.get('*', (_, res) => {
+    res.sendFile(path.join(__dirname, 'dist', 'index.html'))
+  })
+
+const PORT = process.env.PORT || 8080;
+
+app.listen(PORT);
