+++ src/views/Home.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="home">
+    <img alt="Vue logo" src="../assets/logo.png">
+  </div>
+</template>
+
+<script>
+
+</script>
